+++ pkg/apis/management.cattle.io/v3/cluster_types.go
@@ -74,6 +74,7 @@ const (
 	ClusterDriverRKE      = "rancherKubernetesEngine"
 	ClusterDriverK3s      = "k3s"
 	ClusterDriverK3os     = "k3os"
+	ClusterDriverRke2     = "rke2"
 )
 
 // +genclient
@@ -115,6 +116,7 @@ type ClusterSpec struct {
 	Description                         string              `json:"description"`
 	Internal                            bool                `json:"internal" norman:"nocreate,noupdate"`
 	K3sConfig                           *K3sConfig          `json:"k3sConfig,omitempty"`
+	Rke2Config                          *Rke2Config         `json:"rke2Config,omitempty"`
 	ImportedConfig                      *ImportedConfig     `json:"importedConfig,omitempty" norman:"nocreate,noupdate"`
 	GoogleKubernetesEngineConfig        *MapStringInterface `json:"googleKubernetesEngineConfig,omitempty"`
 	AzureKubernetesServiceConfig        *MapStringInterface `json:"azureKubernetesServiceConfig,omitempty"`
